+++ src/Fields/Classification.php
@@ -8,6 +8,11 @@ class Classification extends Subfield implements \JsonSerializable
 {
     use SerializableField;
 
+    const UDC = '080';
+    const DEWEY = '082';
+    const ADD_DEWEY = '082';
+    const OTHER_SCHEME = '084';
+
     /**
      * @var array List of properties to be included when serializing the record using the `toArray()` method.
      */
@@ -23,6 +28,16 @@ class Classification extends Subfield implements \JsonSerializable
             }
         }
         return $out;
+    }
+
+    public function getType()
+    {
+        return $this->getTag();
+    }
+
+    public function getTag()
+    {
+        return $this->field->getTag();
     }
 
     public function getScheme()